Eccentric, aligned-spin waveform model SEOBNRv5EHM
Implementation of the SEOBNRv5EHM waveform model for eccentric, aligned-spin binary black holes in pyseobnr.

New sets of equations have been implemented. Namely, equations for initial conditions of bounded eccentric orbits, equations for the eccentricity corrections to the waveform modes and to the radiation reaction force, as well as auxiliary equations for a set of Keplerian parameters of the orbit.
These equations are employed to evolve the equations of motion of the binary.

A background quasi-circular dynamics with a post-adiabatic evolution (employed in SEOBNRv5HM) is used to: 1) compute the attachment time between the inspiral-plunge and merger-ringdown parts of the eccentric waveforms, and 2) to compute the non-quasicircular coefficients employed to improve the waveforms with respect to quasi-circular numerical-relativity data.

The merger-ringdown prescription is the same as in the quasi-circular model SEOBRv5HM.

Notes:
* The (5,5) mode is not enabled in the model.
* By default, if the starting separation of the binary is less than 10 M, then a set of secular evolution equations is evolved backwards until the separation is larger than 10 M. This has the effect of increasing the available parameter space of the model. This behavior can be configured in the model settings.
* Independently, the user can specify an amount of time to evolve backwards the equations of motion of the binary.

Others:
* "GenerateWaveform" now raises errors for wrong input parameters
* Adding type checks in the input parameters: now boolean are rejected (+unit tests)
* Documentation improvement on "generate_fd_polarizations"
* Fix documentation on how to generate the documentation
* Added tests for the settings on the frequency parameters
* Clean and reorganize the 'containers.pyx' and 'containers.pxd' files
* Fix for very low spins: condition is now coherent with the cython part
  of the code