chore: addressing warnings during build
We have warnings emitted by setuptools such as
_Warning: Package 'pyseobnr.eob.fits.asym_fits.fits_44' is absent from the `packages` configuration.

We have also warnings about content inside "sanity_checks" that is actually not
a subpackage that should be installed.

In order to address these, and considering the bug https://github.com/pypa/setuptools/issues/3260
we have to disable the feature of automatic package data, and specify exclude
paths as well as inclusion patterns. The fit folders now become subpackages
which has to be handled slightly differently in the code.

Other:
* requiring newer versions of setuptools for the new format
  for the license field (PEP 639)
* setuptools is not required for runtime