Fixup: capture exponentiation domain errors in cython
After the change to cpow=True in various cython files, the incorrect domain
calculation were not triggering an exception as before. The exception
was caused by the failing cast of a complex object to double, where
the complex object was created by python when a negative scalar was
exponentiated.

The code now catches nans and raises an exception as before.